refactor: table group package 분리
