+++ src/main/java/kitchenpos/tablegroup/domain/TableGroupValidator.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable.domain;
+package kitchenpos.tablegroup.domain;
 
 import static kitchenpos.common.exception.ExceptionMessage.*;
 
@@ -12,6 +12,7 @@ import org.springframework.util.CollectionUtils;
 import kitchenpos.common.exception.BadRequestException;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.repository.OrderRepository;
+import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.repository.OrderTableRepository;
 
 @Component
