+++ src/main/java/kitchenpos/tablegroup/domain/TableUngroupEvent.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable.domain;
+package kitchenpos.tablegroup.domain;
 
 public class TableUngroupEvent {
