+++ src/main/java/kitchenpos/tablegroup/repository/TableGroupRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.ordertable.repository;
+package kitchenpos.tablegroup.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.ordertable.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroup;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
 }
