+++ src/main/java/kitchenpos/tablegroup/ui/TableGroupRestController.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable.ui;
+package kitchenpos.tablegroup.ui;
 
 import java.net.URI;
 
@@ -9,9 +9,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import kitchenpos.ordertable.application.TableGroupService;
-import kitchenpos.ordertable.dto.TableGroupRequest;
-import kitchenpos.ordertable.dto.TableGroupResponse;
+import kitchenpos.tablegroup.application.TableGroupService;
+import kitchenpos.tablegroup.dto.TableGroupRequest;
+import kitchenpos.tablegroup.dto.TableGroupResponse;
 
 @RestController
 public class TableGroupRestController {
